WU: Turn off Delivery Optimization, post jobs: ask if Restart Computer

+++ CW-CLI.ps1
@@ -2306,6 +2306,7 @@ Function SetupWindowsUpdate {
 	# Declare registry keys locations.
 	$Update1 = "HKLM:\SOFTWARE\Policies\Microsoft\Windows\WindowsUpdate"
 	$Update2 = "HKLM:\SOFTWARE\Policies\Microsoft\Windows\WindowsUpdate\AU"
+	$DeliveryOptimization = "HKU:\S-1-5-20\SOFTWARE\Microsoft\Windows\CurrentVersion\DeliveryOptimization\Settings"
 	if (!(Test-Path $Update1)) {
 		New-Item -Path $Update1 | Out-Null
 		New-Item -Path $Update2 | Out-Null
@@ -2318,10 +2319,12 @@ Function SetupWindowsUpdate {
 	Set-ItemProperty -Path $Update1 -Name DeferFeatureUpdatesPeriodInDays -Type DWord -Value 20
 	Set-ItemProperty -Path $Update2 -Name NoAutoUpdate -Type DWord -Value 1
 	Set-ItemProperty -Path $Update2 -Name NoAutoRebootWithLoggedOnUsers -Type Dword -Value 1
+	Set-ItemProperty -Path $DeliveryOptimization -Name "DownloadMode" -Type DWord -Value 0
 
 	# Print user message; policies applied.
 	$WinUpdatePolicies =@(
 		"Turned off automatic updates"
+		"Turn off Delivery optimization"
 		"Device will no longer auto restart if users are signed in"
 		"Turned off re-installation of apps after Windows Updates"
 		"Delayed quality updates by 4 days"
@@ -2641,11 +2644,14 @@ Function Success {
 	Stop-Process -Name explorer -Force
 	Start-Sleep 3
 	print "CleanWin has finished working."
-	print "This PC is set to restart in 10 seconds, please close this window if you want to halt the restart."
 	print "Thank you for using CleanWin."
-	Start-Sleep 10
+	Write-Warning "This device must restart for all changes to take effect."
+	$restart = ask "Restart this PC now? [y/n]"
+	if ($restart -like "y") {
+		Stop-Transcript
+		Restart-Computer
+	}
 	Stop-Transcript
-	Restart-Computer
 }
 
 # Call the desired functions.
